backend/requirements.txt
absl-py==0.11.0
appdirs==1.4.4
asgiref==3.4.1
astunparse==1.6.3
attrs==21.2.0
audioread==2.1.9
autobahn==21.11.1
Automat==20.2.0
autopep8==1.5.7
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
channels==3.0.0
chardet==4.0.0
click==8.0.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==36.0.0
cycler==0.10.0
daphne==3.0.2
datetime-glob==1.0.8
decorator==4.4.2
Django==3.2.2
django-cors-headers==3.7.0
django-filebrowser==3.14.1
django-filter==2.4.0
django-grappelli==2.15.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
flatbuffers==1.12
gast==0.3.3
google-auth==1.26.1
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.4
h11==0.12.0
h5py==2.10.0
httptools==0.3.0
hyperlink==21.0.0
idna==2.10
imageio==2.9.0
importlib-metadata==3.4.0
incremental==21.3.0
itypes==1.2.0
Jinja2==3.0.3
joblib==1.0.1
Keras-Preprocessing==1.1.2
ketos==2.6.2
kiwisolver==1.3.1
lexery==1.1.1
librosa==0.8.0
llvmlite==0.31.0
Markdown==3.3.3
MarkupSafe==2.0.1
matplotlib==3.3.4
networkx==2.5
numba==0.48.0
numexpr==2.7.2
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.2
Pillow==8.1.0
Pint==0.16.1
pooch==1.3.0
protobuf==3.20.0
psutil==5.8.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0  
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.19.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==6.0
requests==2.25.1
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.1
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.0
service-identity==21.1.0
six==1.15.0
SoundFile==0.10.3.post1
sqlparse==0.4.1
tables==3.6.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2021.2.1
toml==0.10.2
tqdm==4.56.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.7.4.3
uritemplate==4.1.1
urllib3==1.26.3
uvicorn==0.17.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
zope.interface==5.4.0
django-json-widget==1.1.1
